+++ gui-Qt/src/SaldoModel.h
@@ -22,7 +22,7 @@ public:
 	virtual ~SaldoModel();
 
 	void setMemberId(const int aMemberId);
-	int amount() const;
+	float amount() const;
 	void refresh();
 
 	QSqlTableModel* getSaldoTableModel() const;
